Clean up fleet status report sending API
