--- gg-fleet-statusd/src/bus_server.h
@@ -1,10 +1,0 @@
-// aws-greengrass-lite - AWS IoT Greengrass runtime for constrained devices
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-#ifndef GG_FLEET_STATUSD_BUS_SERVER_H
-#define GG_FLEET_STATUSD_BUS_SERVER_H
-
-void gg_fleet_statusd_start_server(void);
-
-#endif
